Changed `splitLines` to merge lines based on whitespace at start Changed import in tests to src instead of dist folder
