+++ tests/parse/attendee.test.ts
@@ -1,6 +1,6 @@
 import { getLine } from "@/lib/parse/utils/line";
 
-import { parseIcsAttendee } from "../..";
+import { parseIcsAttendee } from "@/lib/parse/attendee";
 
 it("Test Ics Attendee Parse", async () => {
   const attendee = `ATTENDEE;MEMBER="mailto:DEV-GROUP@example.com":mailto:joecool@example.com`;
