+++ tests/parse/fixtures/longDescriptionEvent.ics
@@ -1,12 +1,12 @@
 BEGIN:VEVENT
-DESCRIPTION:xx xxxxxnx & xxxxxx\,\n\nxx xxxxxxxxx\, x xx xxnnxxxxnx xxx xxx
+DESCRIPTION:xx xxxxxnx & xxxxxx\,\n\nxx xxxxxxxxx\, nx xxx xxx
    x xxx xn xnxxxxxxxxnxx xxxx xxxxxx xx xxxx xxxxxxx xx xxx nxxx xxxx.\n\nxx
    ’x xxxxxxx: xxxxx://xxx.xxnxxxxn.xxx/xn/xxxxxnxxxxxxxxx/\n\nxxnxxxxxx xx
    xxxxx - xxxxx://xxx.xxnxxxxn.xxx/xn/xxxxxx-xxnxxxxxxx/\n\nxxx xxxxxx - xxx
    xx://xxx.xxxxxn.xxxx/xn/xxxx/xxxxxxx/xxxxxxxx-xxxxxxxxxnx-xnxxnxxx-xxxx-xx
    xx\n\nxxxxxx xxx xx xnxx xx xxx xxxx nxxxx xnx xxxxx xxxx.\n\nxxxnxx\,\n\n
    xxxxx\n\nxxx xxxx xxxn xnxxxxx xx xn xnxxnx xxxxxnx\, xxxxxxx xx xxxxxn xx
-   xxx.\n\nxxxxx xx xxxn xxx xxxxxnx: xxxxx://xxxxx.xxx/xxxxxxxxxx\nxxxxxnx x
+   xxx.\n\nxxxxx xx xxxn xxx xxxxxnx: xxxxx://xxxxx.xxx/\nxxxxxnx x
    x: xxxx xx xxxx\nx xxxxxxx xx xxxxxxxnxxx xx xxx xxx xxx xxxx xxxxxxxx’x
     xxxxxxxxnx xnx xxxxxxxx.\n\nxxxx xn xxxnx xxxx xxxnx:\nxnxxxx xxxxxx xxxx
    -xxxx (x): +x xxx-xxx-xxxx<xxx:+xxxx-xxx-xxxx>\nxxxxxnx xx: xxxx xx xxxx\n
@@ -23,8 +23,8 @@ DESCRIPTION:xx xxxxxnx & xxxxxx\,\n\nxx xxxxxxxxx\, x xx xxnnxxxxnx xxx xxx
 UID:040000008200E00074C5B7101A82E00800000000F0BEC878B1C3D901000000000000000
  010000000DD8309E078039740BC00E43EF7064CA4
 SUMMARY:Some really cool event with long description
-DTSTART;TZID=Pacific Standard Time:20230801T110000
-DTEND;TZID=Pacific Standard Time:20230801T113000
+DTSTART;TZID=America/New_York:20231024T173000
+DTEND;TZID=America/New_York:20231024T180000
 CLASS:PUBLIC
 PRIORITY:5
 DTSTAMP:20231024T204754Z
