+++ tests/parse/recurrenceId.test.ts
@@ -1,6 +1,6 @@
 import { getLine } from "@/lib/parse/utils/line";
 
-import { parseIcsRecurrenceId } from "../..";
+import { parseIcsRecurrenceId } from "@/lib/parse/recurrenceId";
 
 it("Test Ics Recurrence Id Parse", async () => {
   const rId = `RECURRENCE-ID;VALUE=DATE:19960401`;
